+++ src/main/java/net/rom/api/client/ModelTransWrapper.java
@@ -1,0 +1,76 @@
+package net.rom.api.client;
+
+import java.util.List;
+
+import javax.annotation.Nullable;
+import javax.vecmath.Matrix4f;
+
+import org.apache.commons.lang3.tuple.Pair;
+
+import net.minecraft.block.state.IBlockState;
+import net.minecraft.client.renderer.block.model.BakedQuad;
+import net.minecraft.client.renderer.block.model.IBakedModel;
+import net.minecraft.client.renderer.block.model.ItemCameraTransforms;
+import net.minecraft.client.renderer.block.model.ItemCameraTransforms.TransformType;
+import net.minecraft.client.renderer.block.model.ItemOverrideList;
+import net.minecraft.client.renderer.texture.TextureAtlasSprite;
+import net.minecraft.util.EnumFacing;
+import net.minecraftforge.common.model.TRSRTransformation;
+
+abstract public class ModelTransWrapper implements IBakedModel {
+	private final IBakedModel parent;
+
+	public ModelTransWrapper(IBakedModel parent) {
+		this.parent = parent;
+	}
+
+	@Override
+	public boolean isAmbientOcclusion() {
+		return parent.isAmbientOcclusion();
+	}
+
+	@Override
+	public boolean isGui3d() {
+		return parent.isGui3d();
+	}
+
+	@Override
+	public boolean isBuiltInRenderer() {
+		return parent.isBuiltInRenderer();
+	}
+
+	@Override
+	public TextureAtlasSprite getParticleTexture() {
+		return parent.getParticleTexture();
+	}
+
+	@Override
+	@SuppressWarnings("deprecation")
+	public ItemCameraTransforms getItemCameraTransforms() {
+		return parent.getItemCameraTransforms();
+	}
+
+	@Override
+	public List<BakedQuad> getQuads(@Nullable IBlockState state, @Nullable EnumFacing side, long rand) {
+		return parent.getQuads(state, side, rand);
+	}
+
+	@Override
+	public ItemOverrideList getOverrides() {
+		return parent.getOverrides();
+	}
+
+	@Override
+	public Pair<? extends IBakedModel, Matrix4f> handlePerspective(
+			ItemCameraTransforms.TransformType cameraTransformType) {
+		Matrix4f matrix4f = getTransformForPerspective(cameraTransformType);
+
+		if (matrix4f == null) {
+			return Pair.of(this, TRSRTransformation.blockCornerToCenter(TRSRTransformation.identity()).getMatrix());
+		}
+
+		return Pair.of(this, matrix4f);
+	}
+
+	abstract protected Matrix4f getTransformForPerspective(TransformType cameraTransformType);
+}