+++ src/main/java/net/rom/exoplanets/astronomy/yzcetisystem/YzCetiDimensions.java
@@ -1,0 +1,10 @@
+package net.rom.exoplanets.astronomy.yzcetisystem;
+
+import net.minecraft.world.DimensionType;
+
+public class YzCetiDimensions {
+
+	public static DimensionType YZCETIB;
+	public static DimensionType YZCETIC;
+	public static DimensionType YZCETID;
+}