+++ src/main/java/net/rom/exoplanets/astronomy/yzcetisystem/b/worldgen/biome/genlayer/GenLayerYzCetiB.java
@@ -1,0 +1,27 @@
+package net.rom.exoplanets.astronomy.yzcetisystem.b.worldgen.biome.genlayer;
+
+import net.minecraft.world.gen.layer.GenLayer;
+import net.minecraft.world.gen.layer.GenLayerVoronoiZoom;
+import net.minecraft.world.gen.layer.GenLayerZoom;
+
+public abstract class GenLayerYzCetiB extends GenLayer {
+	
+	public GenLayerYzCetiB(long l) {
+		super(l);
+	}
+	
+	public static GenLayer[] createWorld(long l) {
+		GenLayer biomes = new GenLayerYzCetiBBiomes(l);
+		biomes = new GenLayerZoom(1000L, biomes);
+		biomes = new GenLayerZoom(1001L, biomes);
+		biomes = new GenLayerZoom(1002L, biomes);
+		biomes = new GenLayerZoom(1003L, biomes);
+		biomes = new GenLayerZoom(1004L, biomes);
+		biomes = new GenLayerZoom(1005L, biomes);
+		GenLayer genLayerVeronoiZoom = new GenLayerVoronoiZoom(10L, biomes);
+		biomes.initWorldGenSeed(l);
+		genLayerVeronoiZoom.initWorldGenSeed(l);
+		
+		return new GenLayer[] { biomes, genLayerVeronoiZoom };
+	}
+}