+++ src/main/java/net/rom/exoplanets/astronomy/yzcetisystem/c/worldgen/BiomeDecoratorYzCetiC.java
@@ -1,0 +1,18 @@
+package net.rom.exoplanets.astronomy.yzcetisystem.c.worldgen;
+
+import java.util.Random;
+
+import net.minecraft.world.World;
+import net.minecraft.world.biome.Biome;
+import net.rom.api.world.biome.BiomeDecoratorExoBase;
+
+public class BiomeDecoratorYzCetiC extends BiomeDecoratorExoBase {
+
+	public BiomeDecoratorYzCetiC() {
+	}
+
+	@Override
+	protected void generate(Biome biome, World world, Random rand) {
+
+	}
+}