+++ src/main/java/net/rom/exoplanets/conf/SConfigCore.java
@@ -26,7 +26,12 @@ public class SConfigCore {
 	public static boolean enableDebug;
 	public static boolean enableRealism;
 	
+	public static boolean warnBetaBuild;
+	public static int configVersion;
+	
 	private static String CATEGORY_GENERAL_MAIN = "Core Interstellar: Exoplanets Settings";
+	private static String CATEGORY_INTERNAL = "Core Internals !DO NOT CHANGE!";
+
 
 
 	public static void syncConfig(boolean load) {
@@ -40,12 +45,19 @@ public class SConfigCore {
 			config.addCustomCategoryComment(CATEGORY_GENERAL_MAIN, "The Core Settings for Exoplanets");
 			config.setCategoryLanguageKey(CATEGORY_GENERAL_MAIN, "exoplanets.configgui.category.generalmain");
 			config.setCategoryRequiresMcRestart(CATEGORY_GENERAL_MAIN, true);
+			
+			config.addCustomCategoryComment(CATEGORY_INTERNAL, "Core Internals !DO NOT CHANGE!");
+			config.setCategoryLanguageKey(CATEGORY_INTERNAL, "exoplanets.configgui.category.interals");
+			
 
 			enableCheckVersion = config.getBoolean("enableCheckVersion", CATEGORY_GENERAL_MAIN, true, "Enable/Disable Check Version", "exoplanets.configgui.enableCheckVersion");
 			enableOverworldOres = config.getBoolean("enableDebug", CATEGORY_GENERAL_MAIN, false, "Enable/Disable Generation Ores on Overworld", "exoplanets.configgui.enableOverworldOres");
 			enableDebug = config.getBoolean("enableOverworldOres", CATEGORY_GENERAL_MAIN, false, "Enable/Disable Debug mode", "exoplanets.configgui.enableDebug");
 			enableRealism = config.getBoolean("enableRealism", CATEGORY_GENERAL_MAIN, false, "Enabling Realism loads the round & realistic Celestial Body Textures on the Celestial Map", "exoplanets.configgui.enableRealism");
 
+			warnBetaBuild = config.getBoolean("warnOnBetaBuild", CATEGORY_INTERNAL, true, "DO NOT CHANGE");
+			configVersion = config.getInt("configVersion", CATEGORY_INTERNAL, 1, 1, 1, "DO NOT CHANGE");
+			
 			if (config.hasChanged()) {
 				config.save();
 			}
@@ -60,6 +72,7 @@ public class SConfigCore {
 		ConfigCategory configGeneral = config.getCategory(CATEGORY_GENERAL_MAIN);
 		configGeneral.setComment("Core Settings");
 		list.add(new ConfigElement(configGeneral));
+		
 
 
 		return list;