+++ src/main/java/net/rom/exoplanets/init/BlocksRegister.java
@@ -7,60 +7,60 @@ import net.rom.exoplanets.block.BlockYzCetiB;
 import net.rom.exoplanets.block.BlockYzCetiC;
 import net.rom.exoplanets.block.decoration.BlockDeco;
 import net.rom.exoplanets.block.decoration.BlockElectronic;
-import net.rom.exoplanets.block.doors.SpaceDoor;
+import net.rom.exoplanets.block.decoration.BlockSatelliteAntenna;
+import net.rom.exoplanets.block.decoration.BlockStandConsole;
 import net.rom.exoplanets.internal.StellarRegistry;
 
 public class BlocksRegister {
-	
-	public static final BlockYzCetiB yzb_sedimentary = new BlockYzCetiB();
-	public static final BlockYzCetiB yzb_ingneous = new BlockYzCetiB();
-	public static final BlockYzCetiB yzb_metamorphic = new BlockYzCetiB();
-	public static final BlockYzCetiB yzb_smoothsedimentary = new BlockYzCetiB();
-	public static final BlockYzCetiC yzc_sedimentary = new BlockYzCetiC();
-	public static final BlockYzCetiC yzc_ingneous = new BlockYzCetiC();
-	public static final BlockYzCetiC yzc_metamorphic = new BlockYzCetiC();
-	public static final BlockYzCetiC yzc_smoothsedimentary = new BlockYzCetiC();
-	public static final BasicBlock dirt = new BasicBlock(Material.GROUND);
-	public static final BasicBlock tiles = new BasicBlock(Material.ROCK);
-	
-	public static final SpaceDoor spaceDoor = new SpaceDoor();
 
-	public static final BlockElectronic raidController = new BlockElectronic();
-	public static final BlockElectronic lower_raidController = new BlockElectronic();
-	public static final BlockElectronic raidcluster = new BlockElectronic();
-	public static final BlockElectronic dataMonitor = new BlockElectronic();
-	public static final BlockElectronic com = new BlockElectronic();
-	public static final BlockElectronic control = new BlockElectronic();
-	
-	public static final BlockDeco deco = new BlockDeco();
-	
+	// TERRAIN BLOCKS
+	public static final BlockYzCetiB YZB_SEDIMENTARY = new BlockYzCetiB();
+	public static final BlockYzCetiB YZB_INGNEOUS = new BlockYzCetiB();
+	public static final BlockYzCetiB YZB_METAMORPHIC = new BlockYzCetiB();
+	public static final BlockYzCetiB YZB_SMOOTHSEDIMENTARY = new BlockYzCetiB();
+	public static final BlockYzCetiC YZC_SEDIMENTARY = new BlockYzCetiC();
+	public static final BlockYzCetiC YZC_INGNEOUS = new BlockYzCetiC();
+	public static final BlockYzCetiC YZC_METAMORPHIC = new BlockYzCetiC();
+	public static final BlockYzCetiC YZC_SMOOTHSEDIMENTARY = new BlockYzCetiC();
+	public static final BasicBlock DIRT = new BasicBlock(Material.GROUND);
+	public static final BasicBlock TILES = new BasicBlock(Material.ROCK);
+
+	// DECORATION BLOCKS
+	public static final BlockElectronic RAIDCONTROLLER = new BlockElectronic();
+	public static final BlockElectronic LOWER_RAIDCONTROLLER = new BlockElectronic();
+	public static final BlockElectronic RAIDCLUSTER = new BlockElectronic();
+	public static final BlockElectronic DATAMONITOR = new BlockElectronic();
+	public static final BlockElectronic COM = new BlockElectronic();
+	public static final BlockElectronic CONTROL = new BlockElectronic();
+	public static final BlockDeco DECO = new BlockDeco();
+	public static final BlockSatelliteAntenna SATELLITE_ANTENNA = new BlockSatelliteAntenna();
+	public static final BlockStandConsole STAND_CONSOLE = new BlockStandConsole();
+
 	public static void registerAll(StellarRegistry registry) {
-		
-		
-		registry.registerBlock(dirt, "basic1", new ItemBlock(dirt));
-		registry.registerBlock(tiles, "basic2", new ItemBlock(tiles));
-		
-		registry.registerBlock(yzb_sedimentary, "yzb_sedimentary", new ItemBlock(yzb_sedimentary));
-		registry.registerBlock(yzb_ingneous, "yzb_ingneous", new ItemBlock(yzb_ingneous));
-		registry.registerBlock(yzb_metamorphic, "yzb_metamorphic", new ItemBlock(yzb_metamorphic));
-		registry.registerBlock(yzb_smoothsedimentary, "yzb_smoothsedimentary", new ItemBlock(yzb_smoothsedimentary));
-		
-		registry.registerBlock(yzc_sedimentary, "yzc_sedimentary", new ItemBlock(yzc_sedimentary));
-		registry.registerBlock(yzc_ingneous, "yzc_ingneous", new ItemBlock(yzc_ingneous));
-		registry.registerBlock(yzc_metamorphic, "yzc_metamorphic", new ItemBlock(yzc_metamorphic));
-		registry.registerBlock(yzc_smoothsedimentary, "yzc_smoothsedimentary", new ItemBlock(yzc_smoothsedimentary));
-		
-		registry.registerBlock(com, "com_relay", new ItemBlock(com));
-		registry.registerBlock(control, "control", new ItemBlock(control));
-		registry.registerBlock(raidController, "raidcontroller", new ItemBlock(raidController));
-		registry.registerBlock(lower_raidController, "lower_raidcontroller", new ItemBlock(lower_raidController));
-		registry.registerBlock(raidcluster, "raidcluster", new ItemBlock(raidcluster));
-		registry.registerBlock(dataMonitor, "datamonitor", new ItemBlock(dataMonitor));
-		
-		registry.registerBlock(deco, "deco", new BlockDeco.ItemBlock(deco));
-		
-		//registry.registerBlock(spaceDoor, "spacedoor", new ItemBlock(spaceDoor));
-		
+
+		// TERRAIN
+		registry.registerBlock(DIRT, "basic1", new ItemBlock(DIRT));
+		registry.registerBlock(TILES, "basic2", new ItemBlock(TILES));
+		registry.registerBlock(YZB_SEDIMENTARY, "yzb_sedimentary", new ItemBlock(YZB_SEDIMENTARY));
+		registry.registerBlock(YZB_INGNEOUS, "yzb_ingneous", new ItemBlock(YZB_INGNEOUS));
+		registry.registerBlock(YZB_METAMORPHIC, "yzb_metamorphic", new ItemBlock(YZB_METAMORPHIC));
+		registry.registerBlock(YZB_SMOOTHSEDIMENTARY, "yzb_smoothsedimentary", new ItemBlock(YZB_SMOOTHSEDIMENTARY));
+		registry.registerBlock(YZC_SEDIMENTARY, "yzc_sedimentary", new ItemBlock(YZC_SEDIMENTARY));
+		registry.registerBlock(YZC_INGNEOUS, "yzc_ingneous", new ItemBlock(YZC_INGNEOUS));
+		registry.registerBlock(YZC_METAMORPHIC, "yzc_metamorphic", new ItemBlock(YZC_METAMORPHIC));
+		registry.registerBlock(YZC_SMOOTHSEDIMENTARY, "yzc_smoothsedimentary", new ItemBlock(YZC_SMOOTHSEDIMENTARY));
+
+		// DECORATION
+		registry.registerBlock(COM, "com_relay", new ItemBlock(COM));
+		registry.registerBlock(CONTROL, "control", new ItemBlock(CONTROL));
+		registry.registerBlock(RAIDCONTROLLER, "raidcontroller", new ItemBlock(RAIDCONTROLLER));
+		registry.registerBlock(LOWER_RAIDCONTROLLER, "lower_raidcontroller", new ItemBlock(LOWER_RAIDCONTROLLER));
+		registry.registerBlock(RAIDCLUSTER, "raidcluster", new ItemBlock(RAIDCLUSTER));
+		registry.registerBlock(DATAMONITOR, "datamonitor", new ItemBlock(DATAMONITOR));
+		registry.registerBlock(DECO, "deco", new BlockDeco.ItemBlock(DECO));
+		registry.registerBlock(SATELLITE_ANTENNA, "satellite_antenna", new ItemBlock(SATELLITE_ANTENNA));
+		registry.registerBlock(STAND_CONSOLE, "stand_console", new ItemBlock(STAND_CONSOLE));
+
 	}
 
 }